add job to remove stale joins
Fix job logging issue of the text not getting printed out

+++ src/jobs.rs
@@ -1,7 +1,7 @@
 use std::sync::Arc;
 
-use chrono::Utc;
-use serenity::model::id::{GuildId, UserId};
+use chrono::{Duration, Utc};
+use serenity::model::id::{ChannelId, GuildId, UserId};
 use serenity::prelude::*;
 use serenity::utils::MessageBuilder;
 use tracing::{error, info, instrument};
@@ -41,17 +41,122 @@ pub async fn log_system_load(ctx: Arc<Context>) {
     };
 }
 
-/// Remove players from pug if they joined over 6 hours ago
-#[instrument(skip(_ctx))]
-pub async fn clear_out_stale_joins(_ctx: Arc<Context>) {
+/// Remove players from pug if they joined over 6 hours ago.
+///
+/// Since currently there is no enforcing of processing pug commands
+/// in a designated pug channel, when a designated pug channel is not declared,
+/// a DM will be sent to the removed user instead of in a pug channel
+#[instrument(skip(ctx))]
+pub async fn clear_out_stale_joins(ctx: Arc<Context>) {
     let current_time = Utc::now();
     let _formatted_time = current_time.to_rfc2822();
 
-    // _ctx.set_activity(Activity::playing(&_formatted_time)).await;
+    let db_client = {
+        let data = ctx.data.read().await;
+        match data.get::<crate::DbClientRef>() {
+            Some(c) => c.clone(),
+            None => {
+                info!(
+                    "Client for database was not available - skipping stale voice channel removal"
+                );
+                return;
+            }
+        }
+    };
+
+    let guilds = ctx.cache.guilds().await;
+    let mut has_error = false;
+    let mut job_log = MessageBuilder::default();
+    for guild_id in guilds {
+        let client = db_client.clone();
+        let guild_db = client.database(guild_id.0.to_string().as_str());
+
+        match guild_id.name(&ctx).await {
+            Some(name) => {
+                job_log.push(name).push_line(":");
+            }
+            None => {
+                job_log.push(guild_id).push_line(":");
+            }
+        }
+        job_log.push_line("==============");
+
+        match crate::db::read::get_stale_game_mode_joins(guild_db.clone(), Duration::hours(4)).await
+        {
+            Ok(game_mode_joins) => {
+                match crate::db::read::get_pug_channel(guild_db.clone()).await {
+                    Ok(maybe_pug_channel) => {
+                        let mut removed_users: Vec<UserId> = Vec::default();
+                        for join in game_mode_joins {
+                            let u_id = join.player_user_id.parse::<u64>().unwrap();
+                            removed_users.push(UserId(u_id));
+                            tokio::spawn(crate::db::write::remove_player_from_game_mode_queue(
+                                guild_db.clone(),
+                                join.game_mode_label,
+                                u_id,
+                            ));
+                        }
+
+                        // delete joins
+                        match maybe_pug_channel {
+                            Some(pug_channel) => {
+                                // send msg in channel
+                                let mut msg = MessageBuilder::default();
+                                msg.push_line("Players removed due to timeout:");
+                                for user in removed_users {
+                                    msg.push(user).push(" ");
+                                }
+
+                                let _ = ChannelId(pug_channel.channel_id).say(&ctx.http, msg).await;
+                            }
+                            None => {
+                                // send dm to removed users
+                                for user in removed_users {
+                                    match user.create_dm_channel(&ctx.http).await {
+                                        Ok(c) => {
+                                            let _ = c
+                                                .say(
+                                                    &ctx.http,
+                                                    "You have been removed from pug due to timeout",
+                                                )
+                                                .await;
+                                        }
+                                        Err(err) => {
+                                            has_error = true;
+                                            job_log.push_line("Expected joined player to have a valid userid for dm:")
+                                                    .push_line(err);
+                                        }
+                                    };
+                                }
+                            }
+                        }
+                    }
+                    Err(err) => {
+                        has_error = true;
+                        job_log
+                            .push_line("Failed to read pug channel")
+                            .push_line(err);
+                    }
+                }
+            }
+            Err(err) => {
+                has_error = true;
+                job_log
+                    .push_line("Failed to read stale game mode joins")
+                    .push_line(err);
+            }
+        }
+    }
+    let job_log_output = job_log.build();
+    if has_error {
+        error!("{}", job_log_output);
+    } else {
+        info!("{}", job_log_output);
+    }
 }
 
 #[instrument(skip(ctx))]
-pub async fn remove_stale_team_voice_channels(ctx: &Arc<Context>, guilds: &Vec<GuildId>) {
+pub async fn remove_stale_team_voice_channels(ctx: Arc<Context>) {
     // !TODO: make sure to skip deleting a voice channel if it's not empty??
 
     let db_client = {
@@ -67,13 +172,18 @@ pub async fn remove_stale_team_voice_channels(ctx: &Arc<Context>, guilds: &Vec<G
         }
     };
 
+    let guilds = ctx.cache.guilds().await;
+
     for guild_id in guilds {
         let ctx_clone = ctx.clone();
 
         let guild_db = db_client.database(guild_id.0.to_string().as_str());
         let mut has_error = false;
         let mut job_log = MessageBuilder::default();
-        job_log.push_line(guild_id);
+        match guild_id.name(&ctx.cache).await {
+            Some(guild_name) => job_log.push_line(guild_name),
+            None => job_log.push_line(guild_id),
+        };
         match crate::db::read::get_voice_channels_pending_deletion(
             guild_db,
             chrono::Duration::hours(2),
@@ -81,7 +191,15 @@ pub async fn remove_stale_team_voice_channels(ctx: &Arc<Context>, guilds: &Vec<G
         .await
         {
             Ok(team_voice_channels) => {
+                if team_voice_channels.is_empty() {
+                    job_log.push_line("No stale voice channels found");
+                    break;
+                }
                 for channel_set in team_voice_channels {
+                    // gather list of all those channel ids that were either deleted/unkown
+                    // FIXME: make db request to flip booleans to true (mark as deleted)
+                    // for documents where at least 1/3 of its voice channel ids are in the list
+
                     match ctx_clone
                         .http
                         .delete_channel(channel_set.category_id.parse::<u64>().unwrap())
@@ -95,6 +213,7 @@ pub async fn remove_stale_team_voice_channels(ctx: &Arc<Context>, guilds: &Vec<G
                             ));
                         }
                         Err(err) => {
+                            has_error = true;
                             job_log
                                 .push_line("Failed to delete an item:")
                                 .push_line(err);
@@ -114,6 +233,7 @@ pub async fn remove_stale_team_voice_channels(ctx: &Arc<Context>, guilds: &Vec<G
                             ));
                         }
                         Err(err) => {
+                            has_error = true;
                             job_log
                                 .push_line("Failed to delete an item:")
                                 .push_line(err);
@@ -133,6 +253,7 @@ pub async fn remove_stale_team_voice_channels(ctx: &Arc<Context>, guilds: &Vec<G
                             ));
                         }
                         Err(err) => {
+                            has_error = true;
                             job_log
                                 .push_line("Failed to delete an item:")
                                 .push_line(err);
